+++ packages/framer-motion/src/motion/features/layout.ts
@@ -1,0 +1,8 @@
+import { MeasureLayout } from "./layout/MeasureLayout"
+import { FeaturePackages } from "./types"
+
+export const layout: FeaturePackages = {
+    layout: {
+        MeasureLayout,
+    },
+}