+++ packages/framer-motion/src/motion/features/layout/MeasureLayout.tsx
@@ -10,7 +10,8 @@ import { globalProjectionState } from "../../../projection/node/state"
 import { correctBorderRadius } from "../../../projection/styles/scale-border-radius"
 import { correctBoxShadow } from "../../../projection/styles/scale-box-shadow"
 import { addScaleCorrector } from "../../../projection/styles/scale-correction"
-import { FeatureProps } from "../types"
+import { MotionProps } from "../../types"
+import { VisualElement } from "../../../render/VisualElement"
 
 interface MeasureContextProps {
     layoutGroup: LayoutGroupContextProps
@@ -19,9 +20,10 @@ interface MeasureContextProps {
     safeToRemove?: VoidFunction | null
 }
 
-class MeasureLayoutWithContext extends React.Component<
-    FeatureProps & MeasureContextProps
-> {
+type MeasureProps = MotionProps &
+    MeasureContextProps & { visualElement: VisualElement }
+
+class MeasureLayoutWithContext extends React.Component<MeasureProps> {
     /**
      * This only mounts projection nodes for components that
      * need measuring, we might want to do it for all components
@@ -54,7 +56,8 @@ class MeasureLayoutWithContext extends React.Component<
         globalProjectionState.hasEverUpdated = true
     }
 
-    getSnapshotBeforeUpdate(prevProps: FeatureProps & MeasureContextProps) {
+    getSnapshotBeforeUpdate(prevProps: MeasureProps) {
+        console.log("get snapshot before update")
         const { layoutDependency, visualElement, drag, isPresent } = this.props
         const projection = visualElement.projection
 
@@ -135,7 +138,9 @@ class MeasureLayoutWithContext extends React.Component<
     }
 }
 
-export function MeasureLayout(props: FeatureProps) {
+export function MeasureLayout(
+    props: MotionProps & { visualElement: VisualElement }
+) {
     const [isPresent, safeToRemove] = usePresence()
     const layoutGroup = useContext(LayoutGroupContext)